# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "GigConnect Backend API Platform",
    "description": "Clean, systematic, and fully categorized RESTful API documentation for GigConnect services, workers, bookings, admin operations, and background workflows.",
    "version": "1.0.0"
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter JWT Bearer token: Bearer <token>"
    }
  },
  "paths": {
...
    "/api/workers/": {
      "get": {
        "tags": [
          "Workers"
        ],
        "description": "Get nearby available workers with infinite scroll / lazy loading (5 items per batch), category & skills filtering, top rating sort, and Indian currency minimum charge",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "lng",
            "in": "query",
            "description": "User longitude (e.g. 77.3639)",
            "required": true,
            "type": "number"
          },
          {
            "name": "lat",
            "in": "query",
            "description": "User latitude (e.g. 28.6280)",
            "required": true,
            "type": "number"
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by category (e.g. Plumbing)",
            "type": "string"
          },
          {
            "name": "skill",
            "in": "query",
            "description": "Filter by single skill (e.g. Pipe Fitting)",
            "type": "string"
          },
          {
            "name": "skills",
            "in": "query",
            "description": "Filter by multiple skills comma-separated (e.g. Pipe Fitting, Leak Detection)",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sorting order: top_rated (default), nearest, jobs, price_low, price_high",
            "type": "string",
            "default": "top_rated"
          },
          {
            "name": "radiusInKm",
            "in": "query",
            "description": "Search radius in kilometers (default 10)",
            "type": "number",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Infinite scroll offset (0 for first batch, 5 for next batch, etc.)",
            "type": "integer",
            "default": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Batch size limit from .env (default 5)",
            "type": "integer",
            "default": 5
          },
          {
            "name": "page",
            "in": "query",
            "description": "Alternative page number for scroll (1 for first batch, 2 for next, etc.)",
            "type": "integer",
            "default": 1
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Find Nearby Available Workers (Geo-query & Filters)"
      }
    },
...
    "/api/workers/{workerId}": {
      "get": {
        "description": "Get Worker Public Profile & Badges by ID",
        "parameters": [
          {
            "name": "workerId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Workers"
        ],
        "summary": "Get Worker Public Profile & Badges by ID"
      }
    },
...
    "/api/workers/{workerId}/reviews": {
      "get": {
        "description": "Get Customer Reviews for Worker",
        "parameters": [
          {
            "name": "workerId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Workers"
        ],
        "summary": "Get Customer Reviews for Worker"
      }
    },
    "/api/services/": {
      "post": {
        "description": "Create New Service (Uses Cloudinary)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "basePrice": {
                  "example": "any"
                },
                "estimatedTime": {
                  "example": "any"
                },
                "whatsIncluded": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Services"
        ],
        "summary": "Create New Service (Uses Cloudinary)"
      }
    },
    "/api/services/categories": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "Get All Categories & Grouped Services (24h Redis Cached)",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get All Categories & Grouped Services (24h Redis Cached)"
      }
    },
...
    "/api/services/{serviceId}": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "Get Service Details & Pricing by ID",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Service Details & Pricing by ID"
      }
    },
    "/api/bookings/": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "description": "Customer creates a new booking request with scheduling and emergency SOS support",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceId": {
                  "example": "any"
                },
                "workerId": {
                  "example": "any"
                },
                "problemDescription": {
                  "example": "any"
                },
                "problemPhotos": {
                  "example": "any"
                },
                "addressLine": {
                  "example": "any"
                },
                "coordinates": {
                  "example": "any"
                },
                "scheduledTime": {
                  "example": "any"
                },
                "timeSlot": {
                  "example": "any"
                },
                "bookingType": {
                  "example": "any"
                },
                "isEmergency": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Create New Service Booking",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/bookings/{bookingId}": {
      "get": {
        "description": "Get Booking Details by ID",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-device-id",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Bookings"
        ],
        "summary": "Get Booking Details by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/ai-insights": {
      "get": {
        "description": "Get AI Generated Operational Insights",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Get AI Generated Operational Insights",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/analytics": {
      "get": {
        "description": "Get Business & Financial Analytics Overview",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Get Business & Financial Analytics Overview",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings": {
      "get": {
        "description": "List All Bookings across Platform (Filtered by Status & Date)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List All Bookings across Platform (Filtered by Status & Date)",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings/{bookingId}/assign": {
      "patch": {
        "description": "Manually Reassign Booking to Different Worker",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workerId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Manually Reassign Booking to Different Worker",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings/{bookingId}/status": {
      "patch": {
        "description": "Update Booking Lifecycle Status",
        "parameters": [
          {
            "name": "bookingId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "any"
                },
                "scheduledTime": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Update Booking Lifecycle Status",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings/{id}": {
      "get": {
        "description": "Get Booking Details by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Get Booking Details by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings/{id}/assign": {
      "patch": {
        "description": "Manually Reassign Booking to Different Worker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "workerId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Manually Reassign Booking to Different Worker",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/bookings/{id}/status": {
      "patch": {
        "description": "Update Booking Lifecycle Status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "any"
                },
                "scheduledTime": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Update Booking Lifecycle Status",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/admin/categories": {
      "get": {
        "description": "Get All Service Categories for Admin",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Get All Service Categories for Admin",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Create New Category with Single Image Upload & Instant Redis Push",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "image": {
                  "example": "any"
                },
                "title": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "basePrice": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "estimatedTime": {
                  "example": "any"
                },
                "whatsIncluded": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "Create New Category with Single Image Upload & Instant Redis Push",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/admin/cooperative/members": {
      "get": {
        "description": "List Cooperative Enrolled Worker Members",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "societyId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Admin"
        ],
        "summary": "List Cooperative Enrolled Worker Members",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}